% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_significance.R
\name{read_sequence}
\alias{read_sequence}
\title{Statistical significance}
\usage{
read_sequence(seq1, seq2, seq.type)
}
\arguments{
\item{seq1}{First sequence as .fasta file}

\item{seq2}{Second sequence as .fasta file}

\item{seq.type}{Type of sequences (DNA/RNA/Protein)}
}
\value{
A list of the two sequences converted to given type
}
\description{
Calculates probability of an alignment score to have been obtained randomly
}
